[pkg/datadog] Move feature gates into their own module
